test: use is_relation_joined to ensure status
